added Filter User By Id #6
